BC
Started experiment to prepare Jupyter notebook about Titanic in Rodeo. That IDE resembles RStudio a lot. So far, it looks promising -- it rides on the Anaconda distribution without any setup. I do not like the composing of notebooks -- changing of cells, etc. Once the contents of a notebook  are finished, I put them in.